Address PR review: uint64_t helper params, consistent assert placement, style
- Change kv_object_ram / kv_index_object_ram parameters from uint32_t
  to uint64_t to eliminate implicit narrowing from size_t and a
  theoretical uint32 addition wrap.
- Move static_asserts to after value definition consistently across
  kv_table_objects.hpp, permission_link_object.hpp (matching the
  permission_object.hpp and authority.hpp placement).
- Condense key_weight comment to one line.
- Replace Unicode separators and em-dashes with ASCII equivalents.
- Convert C-style (int64_t) casts to static_cast<int64_t>() in tests.
- Avoid redundant disk read in setcode_update_ram_billing.

+++ libraries/chain/apply_context.cpp
@@ -18,13 +18,13 @@ namespace sysio::chain {
 
 // RAM billing formulas for kv primary rows and secondary index rows. Variable
 // key/value bytes plus a fixed-field overhead defined by billable_size<>.
-static inline int64_t kv_object_ram(uint32_t key_size, uint32_t value_size) {
+static inline int64_t kv_object_ram(uint64_t key_size, uint64_t value_size) {
    return static_cast<int64_t>(key_size + value_size + config::billable_size_v<kv_object>);
 }
 static inline int64_t kv_object_ram(const kv_object& o) {
    return kv_object_ram(o.key.size(), o.value.size());
 }
-static inline int64_t kv_index_object_ram(uint32_t sec_key_size, uint32_t pri_key_size) {
+static inline int64_t kv_index_object_ram(uint64_t sec_key_size, uint64_t pri_key_size) {
    return static_cast<int64_t>(sec_key_size + pri_key_size + config::billable_size_v<kv_index_object>);
 }
 static inline int64_t kv_index_object_ram(const kv_index_object& o) {
